Adding hook for fetching tags
